refine complie configuration
