+++ .gitignore
@@ -11,7 +11,6 @@
 
 internal/core/output/*
 internal/core/build/*
-internal/kv/rocksdb/cwrapper/output/*
 **/.idea/*
 internal/msgstream/pulsarms/client-cpp/build/
 internal/msgstream/pulsarms/client-cpp/build/*
@@ -56,7 +55,6 @@ lib/
 .DS_Store
 *.sw[po]
 cwrapper_build
-**/cwrapper_rocksdb_build/
 **/.clangd/*
 **/compile_commands.json
 **/.lint
