+++ internal/core/src/CMakeLists.txt
@@ -28,3 +28,4 @@ add_subdirectory( common )
 add_subdirectory( indexbuilder )
 add_subdirectory( config )
 add_subdirectory( index )
+add_subdirectory( storage )
