+++ internal/core/src/indexbuilder/milvus_indexbuilder.pc.in
@@ -1,0 +1,9 @@
+libdir=@CMAKE_INSTALL_FULL_LIBDIR@
+includedir=@CMAKE_INSTALL_FULL_INCLUDEDIR@
+
+Name: Milvus IndexBuilder
+Description: IndexBuilder modules for Milvus
+Version: @MILVUS_VERSION@
+
+Libs: -L${libdir} -lmilvus_indexbuilder
+Cflags: -I${includedir}
