+++ internal/core/src/storage/milvus_storage.pc.in
@@ -1,0 +1,10 @@
+
+libdir=@CMAKE_INSTALL_FULL_LIBDIR@
+includedir=@CMAKE_INSTALL_FULL_INCLUDEDIR@
+
+Name: Milvus Storage
+Description: Storage wrapper for Milvus
+Version: @MILVUS_VERSION@
+
+Libs: -L${libdir} -lmilvus_storage
+Cflags: -I${includedir}
