+++ internal/core/unittest/CMakeLists.txt
@@ -29,6 +29,7 @@ set(MILVUS_TEST_FILES
         test_index_c_api.cpp
         test_index_wrapper.cpp
         test_init.cpp
+        test_parquet_c.cpp
         test_plan_proto.cpp
         test_query.cpp
         test_reduce.cpp
@@ -87,6 +88,7 @@ target_link_libraries(all_tests
         milvus_indexbuilder
         milvus_index
         milvus_log
+        milvus_storage
         pthread
         )
