--- internal/storage/cwrapper/.gitignore
@@ -1,4 +1,0 @@
-output
-cmake-build-debug
-.idea
-cmake_build
