+++ internal/storage/payload.go
@@ -17,13 +17,10 @@
 package storage
 
 /*
-#cgo CFLAGS: -I${SRCDIR}/cwrapper
+#cgo pkg-config: milvus_storage
 
-#cgo linux LDFLAGS: -L${SRCDIR}/cwrapper/output/lib -L${SRCDIR}/cwrapper/output/lib64 -lwrapper -lparquet -larrow -larrow_bundled_dependencies -lstdc++ -lm
-#cgo darwin LDFLAGS: -L${SRCDIR}/cwrapper/output/lib -lwrapper -lparquet -larrow -larrow_bundled_dependencies -lstdc++ -lm
-#cgo windows LDFLAGS: -L${SRCDIR}/cwrapper/output/lib -lwrapper -lparquet -larrow -lstdc++
 #include <stdlib.h>
-#include "ParquetWrapper.h"
+#include "storage/parquet_c.h"
 */
 import "C"
 import (
